Add files via upload
A new version with added with the following changes. An improved method for displaying lists. Can now display both footnotes and endnotes together with links between the them and the reference number in the text. Will also display a Table of Contents along with the bookmarks which link to sections of the document. Also a few bug fixes